+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/java/Java.java
@@ -20,9 +20,9 @@ import static java.util.Collections.unmodifiableSet;
 
 import java.util.Set;
 
-import com.diffplug.spotless.maven.FormatterFactory;
+import com.diffplug.spotless.maven.generic.Format;
 
-public class Java extends FormatterFactory {
+public class Java extends Format {
 	private static final Set<String> DEFAULT_INCLUDES = unmodifiableSet(newHashSet("src/main/java/**/*.java",
 			"src/test/java/**/*.java"));
 	private static final String LICENSE_HEADER_DELIMITER = "package ";